Fix #5 java.net.UnknownHostException by preventing DTD resolution
Prior to this commit XML Formatter would require a DTD to be
available. This would cause failures if formatting was attempted
without a network connection.

This commit replaces the entity resolver with a no-op. This means the
XML files will not be validated against a DTD. This is a reasonable
compromise given XML validation is not in the scope of this
plugin. As a further bonus, the plugin will now execute faster without
the network fetch latencies.