[pom] Fix maven setup and scopes, update all plugin versions
note: know there is a parent pom, this is to show where to go, expect parent pom is updated and then downstream just is up to date.  With this, license plugin still has maven issues for 3.9 that are not released, same with jacoco, and surefire does but I just did not update here since no direct reference.