[fix] Correct issue 36 on keep new line
The data may have something like 'someValue\n' because \n was not tokenized as normal.  If that is allowed to skip, it will not print 'someValue' at all.  Thus this needs to check that it has more tokens before it decides to continue in order to retain the data and still print new line at the end.  Or this in this example not given its just 1.